refac: move `EndianAutoReset` to `base/`
